refactor: move Action class in its own module core.action
Needed to avoid circular imports.
