+++ README.md
@@ -82,7 +82,8 @@ programmatically:
 
 ```python
 from pddl.logic import Predicate, constants, variables
-from pddl.core import Domain, Problem, Action
+from pddl.core import Domain, Problem
+from pddl.action import Action
 from pddl.formatter import domain_to_string, problem_to_string
 from pddl.requirements import Requirements
