+++ tests/test_action.py
@@ -11,7 +11,7 @@
 #
 """This module contains tests for a PDDL action."""
 
-from pddl.core import Action
+from pddl.action import Action
 from pddl.logic import Predicate, Variable, variables
 from pddl.logic.base import ExistsCondition, ForallCondition, Imply, OneOf
